+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - paging when displayed inside a form.
 
+### Added
+
+- possibility to customize back/next first/last page navigation components.
+
 ### Removed
 
 - remove cypress declarations